Splitted TaperConfigPage into create/edit/confirm pages. Working on fixing routing
