+++ src/pages/TaperConfiguration/ConfirmTaperConfiguration.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+const ConfirmTaperConfiguration = () => {
+  return <div>ConfirmTaperConfiguration</div>;
+};
+
+export default ConfirmTaperConfiguration;
