+++ src/pages/TaperConfiguration/index.ts
@@ -1,0 +1,8 @@
+import TaperConfigurationPage from './TaperConfigurationPage';
+import CreateTaperConfiguration from './CreateTaperConfiguration';
+import EditTaperConfiguration from './EditTaperConfiguration';
+import ConfirmTaperConfiguration from './ConfirmTaperConfiguration';
+
+export {
+  TaperConfigurationPage, CreateTaperConfiguration, EditTaperConfiguration, ConfirmTaperConfiguration,
+};
